# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Taxi Ride Matching API
  description: |
    REST API for creating users and ride requests, discovering route-compatible
    travelers, completing two-sided match acceptance, and messaging in the private
    chat created for an accepted match.

    Typical workflow:
    1. Create two users.
    2. Create one future ride for each user. Matching runs when a ride is created.
    3. Retrieve a ride's potential matches.
    4. Both participating users accept the same match.
    5. Use the returned chat ID to send and retrieve messages.

    Authentication is not implemented in this MVP. Acting user IDs are supplied in
    request bodies and must belong to the relevant ride or chat.
  contact:
    name: Taxi API maintainers
  license:
    name: Private project
    identifier: LicenseRef-Proprietary
  version: 1.0.0-MVP
...
servers:
- url: http://localhost:8080
  description: Local development
...
paths:
...
  /api/users:
    get:
      tags:
      - Users
      summary: Find a user by email
      description: Returns an existing user by email so the same account can create
        additional ride requests without being registered again.
      operationId: getByEmail
      parameters:
      - name: email
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: User found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "404":
          description: No user has this email
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
    post:
      tags:
      - Users
      summary: Create a user
      description: Creates a user with a unique email address. The returned UUID is
        used when creating rides and acting in match/chat workflows.
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
        required: true
      responses:
        "201":
          description: User created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Validation failed
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "409":
          description: Email already exists
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
...
  /api/users/{id}:
    get:
      tags:
      - Users
      summary: Get a user
      description: Returns one user by UUID.
      operationId: get_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: User found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
  /api/rides/{rideId}/matches:
    get:
      tags:
      - Matches
      summary: List potential matches for a ride
      description: "Returns persisted matches ordered by score. An empty array means\
        \ no candidate passed user, time, proximity, destination, overlap, and minimum-score\
        \ checks."
      operationId: getMatches
      parameters:
      - name: rideId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Match list returned
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RideMatchResponse"
        "404":
          description: Ride not found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RideMatchResponse"
...
components:
  schemas:
...
    CreateUserRequest:
      type: object
      description: Information required to register an MVP user
      properties:
        firstName:
          type: string
          example: Nithin
          maxLength: 100
          minLength: 0
        lastName:
          type: string
          example: User
          maxLength: 100
          minLength: 0
        email:
          type: string
          format: email
          example: nithin@example.com
          maxLength: 320
          minLength: 0
        phoneNumber:
          type: string
          example: +1 312-555-0101
          pattern: "^$|^[+0-9(). -]{7,30}$"
      required:
      - email
      - firstName
      - lastName
    UserResponse:
      type: object
      description: Registered platform user
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    RideMatchResponse:
      type: object
      description: Compatibility evaluation for two ride requests. Scores range from
        0 to 1; distances are meters.
      properties:
        id:
          type: string
          format: uuid
        rideRequest1Id:
          type: string
          format: uuid
        rideRequest2Id:
          type: string
          format: uuid
        matchScore:
          type: number
          format: double
        pickupDistanceMeters:
          type: number
          format: double
        destinationDistanceMeters:
          type: number
          format: double
        routeSimilarity:
          type: number
          format: double
        departureTimeDifferenceSeconds:
          type: integer
          format: int64
        status:
          type: string
          enum:
          - PENDING
          - ACCEPTED
          - REJECTED
          - CANCELLED
        rideRequest1AcceptedAt:
          type: string
          format: date-time
        rideRequest2AcceptedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...